% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.directconnect_operations.R
\name{allocate_connection_on_interconnect}
\alias{allocate_connection_on_interconnect}
\title{Deprecated}
\usage{
allocate_connection_on_interconnect(bandwidth, connectionName,
  ownerAccount, interconnectId, vlan)
}
\arguments{
\item{bandwidth}{[required] The bandwidth of the connection, in Mbps. The possible values are 50Mbps, 100Mbps, 200Mbps, 300Mbps, 400Mbps, and 500Mbps.}

\item{connectionName}{[required] The name of the provisioned connection.}

\item{ownerAccount}{[required] The ID of the AWS account of the customer for whom the connection will be provisioned.}

\item{interconnectId}{[required] The ID of the interconnect on which the connection will be provisioned. For example, dxcon-456abc78.}

\item{vlan}{[required] The dedicated VLAN provisioned to the connection.}
}
\description{
Deprecated. Use AllocateHostedConnection instead.
}
\details{
Creates a hosted connection on an interconnect.

Allocates a VLAN number and a specified amount of bandwidth for use by a hosted connection on the specified interconnect.

Intended for use by AWS Direct Connect partners only.
}
\section{Accepted Parameters}{
\preformatted{allocate_connection_on_interconnect(
  bandwidth = "string",
  connectionName = "string",
  ownerAccount = "string",
  interconnectId = "string",
  vlan = 123
)
}
}

