% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.directconnect_operations.R
\name{allocate_hosted_connection}
\alias{allocate_hosted_connection}
\title{Creates a hosted connection on the specified interconnect or a link aggregation group (LAG)}
\usage{
allocate_hosted_connection(connectionId, ownerAccount, bandwidth,
  connectionName, vlan)
}
\arguments{
\item{connectionId}{[required] The ID of the interconnect or LAG.}

\item{ownerAccount}{[required] The ID of the AWS account ID of the customer for the connection.}

\item{bandwidth}{[required] The bandwidth of the hosted connection, in Mbps. The possible values are 50Mbps, 100Mbps, 200Mbps, 300Mbps, 400Mbps, and 500Mbps.}

\item{connectionName}{[required] The name of the hosted connection.}

\item{vlan}{[required] The dedicated VLAN provisioned to the hosted connection.}
}
\description{
Creates a hosted connection on the specified interconnect or a link aggregation group (LAG).
}
\details{
Allocates a VLAN number and a specified amount of bandwidth for use by a hosted connection on the specified interconnect or LAG.

Intended for use by AWS Direct Connect partners only.
}
\section{Accepted Parameters}{
\preformatted{allocate_hosted_connection(
  connectionId = "string",
  ownerAccount = "string",
  bandwidth = "string",
  connectionName = "string",
  vlan = 123
)
}
}

