% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.directconnect_operations.R
\name{allocate_private_virtual_interface}
\alias{allocate_private_virtual_interface}
\title{Provisions a private virtual interface to be owned by the specified AWS account}
\usage{
allocate_private_virtual_interface(connectionId, ownerAccount,
  newPrivateVirtualInterfaceAllocation)
}
\arguments{
\item{connectionId}{The ID of the connection on which the private virtual interface is provisioned.}

\item{ownerAccount}{The ID of the AWS account that owns the virtual private interface.}

\item{newPrivateVirtualInterfaceAllocation}{Information about the private virtual interface.}
}
\description{
Provisions a private virtual interface to be owned by the specified AWS account.
}
\details{
Virtual interfaces created using this action must be confirmed by the owner using ConfirmPrivateVirtualInterface. Until then, the virtual interface is in the \code{Confirming} state and is not available to handle traffic.
}
\examples{

}
