% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.directconnect_operations.R
\name{confirm_connection}
\alias{confirm_connection}
\title{Confirms the creation of the specified hosted connection on an interconnect}
\usage{
confirm_connection(
  connectionId = "string"
)
}
\arguments{
\item{connectionId}{[required] The ID of the hosted connection.}
}
\description{
Confirms the creation of the specified hosted connection on an interconnect.
}
\details{
Upon creation, the hosted connection is initially in the \code{Ordering} state, and remains in this state until the owner confirms creation of the hosted connection.
}
\examples{

}
