% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.directconnect_operations.R
\name{confirm_public_virtual_interface}
\alias{confirm_public_virtual_interface}
\title{Accepts ownership of a public virtual interface created by another AWS account}
\usage{
confirm_public_virtual_interface(virtualInterfaceId)
}
\arguments{
\item{virtualInterfaceId}{The ID of the virtual interface.}
}
\description{
Accepts ownership of a public virtual interface created by another AWS account.
}
\details{
After the virtual interface owner makes this call, the specified virtual interface is created and made available to handle traffic.
}
\examples{

}
