% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.directconnect_operations.R
\name{create_connection}
\alias{create_connection}
\title{Creates a connection between a customer network and a specific AWS Direct Connect location}
\usage{
create_connection(location, bandwidth, connectionName, lagId = NULL)
}
\arguments{
\item{location}{[required] The location of the connection.}

\item{bandwidth}{[required] The bandwidth of the connection.}

\item{connectionName}{[required] The name of the connection.}

\item{lagId}{The ID of the LAG.}
}
\description{
Creates a connection between a customer network and a specific AWS Direct Connect location.
}
\details{
A connection links your internal network to an AWS Direct Connect location over a standard Ethernet fiber-optic cable. One end of the cable is connected to your router, the other to an AWS Direct Connect router.

To find the locations for your Region, use DescribeLocations.

You can automatically add the new connection to a link aggregation group (LAG) by specifying a LAG ID in the request. This ensures that the new connection is allocated on the same AWS Direct Connect endpoint that hosts the specified LAG. If there are no available ports on the endpoint, the request fails and no connection is created.
}
\section{Accepted Parameters}{
\preformatted{create_connection(
  location = "string",
  bandwidth = "string",
  connectionName = "string",
  lagId = "string"
)
}
}

