% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.directconnect_operations.R
\name{create_direct_connect_gateway_association}
\alias{create_direct_connect_gateway_association}
\title{Creates an association between a Direct Connect gateway and a virtual private gateway}
\usage{
create_direct_connect_gateway_association(
  directConnectGatewayId = "string",
  virtualGatewayId = "string"
)
}
\arguments{
\item{directConnectGatewayId}{[required] The ID of the Direct Connect gateway.}

\item{virtualGatewayId}{[required] The ID of the virtual private gateway.}
}
\description{
Creates an association between a Direct Connect gateway and a virtual private gateway. The virtual private gateway must be attached to a VPC and must not be associated with another Direct Connect gateway.
}
\examples{

}
