% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.directconnect_operations.R
\name{create_interconnect}
\alias{create_interconnect}
\title{Creates an interconnect between an AWS Direct Connect partner's network and a specific AWS Direct Connect location}
\usage{
create_interconnect(interconnectName, bandwidth, location, lagId = NULL)
}
\arguments{
\item{interconnectName}{[required] The name of the interconnect.}

\item{bandwidth}{[required] The port bandwidth, in Gbps. The possible values are 1 and 10.}

\item{location}{[required] The location of the interconnect.}

\item{lagId}{The ID of the LAG.}
}
\description{
Creates an interconnect between an AWS Direct Connect partner's network and a specific AWS Direct Connect location.
}
\details{
An interconnect is a connection which is capable of hosting other connections. The partner can use an interconnect to provide sub-1Gbps AWS Direct Connect service to tier 2 customers who do not have their own connections. Like a standard connection, an interconnect links the partner's network to an AWS Direct Connect location over a standard Ethernet fiber-optic cable. One end is connected to the partner's router, the other to an AWS Direct Connect router.

You can automatically add the new interconnect to a link aggregation group (LAG) by specifying a LAG ID in the request. This ensures that the new interconnect is allocated on the same AWS Direct Connect endpoint that hosts the specified LAG. If there are no available ports on the endpoint, the request fails and no interconnect is created.

For each end customer, the AWS Direct Connect partner provisions a connection on their interconnect by calling AllocateConnectionOnInterconnect. The end customer can then connect to AWS resources by creating a virtual interface on their connection, using the VLAN assigned to them by the partner.

Intended for use by AWS Direct Connect partners only.
}
\section{Accepted Parameters}{
\preformatted{create_interconnect(
  interconnectName = "string",
  bandwidth = "string",
  location = "string",
  lagId = "string"
)
}
}

