% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.directconnect_operations.R
\name{delete_connection}
\alias{delete_connection}
\title{Deletes the specified connection}
\usage{
delete_connection(connectionId)
}
\arguments{
\item{connectionId}{[required] The ID of the connection.}
}
\description{
Deletes the specified connection.
}
\details{
Deleting a connection only stops the AWS Direct Connect port hour and data transfer charges. If you are partnering with any third parties to connect with the AWS Direct Connect location, you must cancel your service with them separately.
}
\section{Accepted Parameters}{
\preformatted{delete_connection(
  connectionId = "string"
)
}
}

