% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.directconnect_operations.R
\name{delete_direct_connect_gateway_association}
\alias{delete_direct_connect_gateway_association}
\title{Deletes the association between the specified Direct Connect gateway and virtual private gateway}
\usage{
delete_direct_connect_gateway_association(
  directConnectGatewayId = "string",
  virtualGatewayId = "string"
)
}
\arguments{
\item{directConnectGatewayId}{[required] The ID of the Direct Connect gateway.}

\item{virtualGatewayId}{[required] The ID of the virtual private gateway.}
}
\description{
Deletes the association between the specified Direct Connect gateway and virtual private gateway.
}
\examples{

}
