% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.directconnect_operations.R
\name{delete_interconnect}
\alias{delete_interconnect}
\title{Deletes the specified interconnect}
\usage{
delete_interconnect(interconnectId)
}
\arguments{
\item{interconnectId}{[required] The ID of the interconnect.}
}
\description{
Deletes the specified interconnect.
}
\details{
Intended for use by AWS Direct Connect partners only.
}
\section{Accepted Parameters}{
\preformatted{delete_interconnect(
  interconnectId = "string"
)
}
}

