% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.directconnect_operations.R
\name{delete_virtual_interface}
\alias{delete_virtual_interface}
\title{Deletes a virtual interface}
\usage{
delete_virtual_interface(virtualInterfaceId)
}
\arguments{
\item{virtualInterfaceId}{[required] The ID of the virtual interface.}
}
\description{
Deletes a virtual interface.
}
\section{Accepted Parameters}{
\preformatted{delete_virtual_interface(
  virtualInterfaceId = "string"
)
}
}

