% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.directconnect_operations.R
\name{describe_connection_loa}
\alias{describe_connection_loa}
\title{Deprecated}
\usage{
describe_connection_loa(
  connectionId = "string",
  providerName = "string",
  loaContentType = "application/pdf"
)
}
\arguments{
\item{connectionId}{[required] The ID of the connection.}

\item{providerName}{The name of the APN partner or service provider who establishes connectivity on your behalf. If you specify this parameter, the LOA-CFA lists the provider name alongside your company name as the requester of the cross connect.}

\item{loaContentType}{The standard media type for the LOA-CFA document. The only supported value is application/pdf.}
}
\description{
Deprecated. Use DescribeLoa instead.
}
\details{
Gets the LOA-CFA for a connection.

The Letter of Authorization - Connecting Facility Assignment (LOA-CFA) is a document that your APN partner or service provider uses when establishing your cross connect to AWS at the colocation facility. For more information, see \href{http://docs.aws.amazon.com/directconnect/latest/UserGuide/Colocation.html}{Requesting Cross Connects at AWS Direct Connect Locations} in the \emph{AWS Direct Connect User Guide}.
}
\examples{

}
