% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.directconnect_operations.R
\name{describe_connections}
\alias{describe_connections}
\title{Displays the specified connection or all connections in this Region}
\usage{
describe_connections(
  connectionId = "string"
)
}
\arguments{
\item{connectionId}{The ID of the connection.}
}
\description{
Displays the specified connection or all connections in this Region.
}
\examples{

}
