% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.directconnect_operations.R
\name{describe_connections_on_interconnect}
\alias{describe_connections_on_interconnect}
\title{Deprecated}
\usage{
describe_connections_on_interconnect(interconnectId)
}
\arguments{
\item{interconnectId}{The ID of the interconnect.}
}
\description{
Deprecated. Use DescribeHostedConnections instead.
}
\details{
Lists the connections that have been provisioned on the specified interconnect.

Intended for use by AWS Direct Connect partners only.
}
\examples{

}
