% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.directconnect_operations.R
\name{describe_direct_connect_gateway_associations}
\alias{describe_direct_connect_gateway_associations}
\title{Lists the associations between your Direct Connect gateways and virtual private gateways}
\usage{
describe_direct_connect_gateway_associations(
  directConnectGatewayId = "string",
  virtualGatewayId = "string",
  maxResults = 123,
  nextToken = "string"
)
}
\arguments{
\item{directConnectGatewayId}{The ID of the Direct Connect gateway.}

\item{virtualGatewayId}{The ID of the virtual private gateway.}

\item{maxResults}{The maximum number of associations to return per page.}

\item{nextToken}{The token provided in the previous call to retrieve the next page.}
}
\description{
Lists the associations between your Direct Connect gateways and virtual private gateways. You must specify a Direct Connect gateway, a virtual private gateway, or both. If you specify a Direct Connect gateway, the response contains all virtual private gateways associated with the Direct Connect gateway. If you specify a virtual private gateway, the response contains all Direct Connect gateways associated with the virtual private gateway. If you specify both, the response contains the association between the Direct Connect gateway and the virtual private gateway.
}
\examples{

}
