% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.directconnect_operations.R
\name{describe_hosted_connections}
\alias{describe_hosted_connections}
\title{Lists the hosted connections that have been provisioned on the specified interconnect or link aggregation group (LAG)}
\usage{
describe_hosted_connections(connectionId)
}
\arguments{
\item{connectionId}{[required] The ID of the interconnect or LAG.}
}
\description{
Lists the hosted connections that have been provisioned on the specified interconnect or link aggregation group (LAG).
}
\details{
Intended for use by AWS Direct Connect partners only.
}
\section{Accepted Parameters}{
\preformatted{describe_hosted_connections(
  connectionId = "string"
)
}
}

