% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.directconnect_operations.R
\name{describe_interconnect_loa}
\alias{describe_interconnect_loa}
\title{Deprecated}
\usage{
describe_interconnect_loa(interconnectId, providerName = NULL,
  loaContentType = NULL)
}
\arguments{
\item{interconnectId}{[required] The ID of the interconnect.}

\item{providerName}{The name of the service provider who establishes connectivity on your behalf. If you supply this parameter, the LOA-CFA lists the provider name alongside your company name as the requester of the cross connect.}

\item{loaContentType}{The standard media type for the LOA-CFA document. The only supported value is application/pdf.}
}
\description{
Deprecated. Use DescribeLoa instead.
}
\details{
Gets the LOA-CFA for the specified interconnect.

The Letter of Authorization - Connecting Facility Assignment (LOA-CFA) is a document that is used when establishing your cross connect to AWS at the colocation facility. For more information, see \href{http://docs.aws.amazon.com/directconnect/latest/UserGuide/Colocation.html}{Requesting Cross Connects at AWS Direct Connect Locations} in the \emph{AWS Direct Connect User Guide}.
}
\section{Accepted Parameters}{
\preformatted{describe_interconnect_loa(
  interconnectId = "string",
  providerName = "string",
  loaContentType = "application/pdf"
)
}
}

