% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.directconnect_operations.R
\name{describe_interconnects}
\alias{describe_interconnects}
\title{Lists the interconnects owned by the AWS account or only the specified interconnect}
\usage{
describe_interconnects(
  interconnectId = "string"
)
}
\arguments{
\item{interconnectId}{The ID of the interconnect.}
}
\description{
Lists the interconnects owned by the AWS account or only the specified interconnect.
}
\examples{

}
