% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.directconnect_operations.R
\name{describe_lags}
\alias{describe_lags}
\title{Describes all your link aggregation groups (LAG) or the specified LAG}
\usage{
describe_lags(
  lagId = "string"
)
}
\arguments{
\item{lagId}{The ID of the LAG.}
}
\description{
Describes all your link aggregation groups (LAG) or the specified LAG.
}
\examples{

}
