% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.directconnect_operations.R
\name{describe_locations}
\alias{describe_locations}
\title{Lists the AWS Direct Connect locations in the current AWS Region}
\usage{
describe_locations()
}
\description{
Lists the AWS Direct Connect locations in the current AWS Region. These are the locations that can be selected when calling CreateConnection or CreateInterconnect.
}
\examples{

}
