% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.directconnect_operations.R
\name{describe_tags}
\alias{describe_tags}
\title{Describes the tags associated with the specified AWS Direct Connect resources}
\usage{
describe_tags(
  resourceArns = list(
    "string"
  )
)
}
\arguments{
\item{resourceArns}{[required] The Amazon Resource Names (ARNs) of the resources.}
}
\description{
Describes the tags associated with the specified AWS Direct Connect resources.
}
\examples{

}
