% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.directconnect_operations.R
\name{describe_virtual_interfaces}
\alias{describe_virtual_interfaces}
\title{Displays all virtual interfaces for an AWS account}
\usage{
describe_virtual_interfaces(connectionId = NULL, virtualInterfaceId = NULL)
}
\arguments{
\item{connectionId}{The ID of the connection.}

\item{virtualInterfaceId}{The ID of the virtual interface.}
}
\description{
Displays all virtual interfaces for an AWS account. Virtual interfaces deleted fewer than 15 minutes before you make the request are also returned. If you specify a connection ID, only the virtual interfaces associated with the connection are returned. If you specify a virtual interface ID, then only a single virtual interface is returned.
}
\details{
A virtual interface (VLAN) transmits the traffic between the AWS Direct Connect location and the customer network.
}
\examples{

}
