% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.directconnect_package.R
\docType{package}
\name{paws.directconnect-package}
\alias{paws.directconnect}
\alias{paws.directconnect-package}
\title{paws.directconnect: AWS Direct Connect}
\description{
AWS Direct Connect links your internal network to
    an AWS Direct Connect location over a standard Ethernet fiber-optic
    cable. One end of the cable is connected to your router, the other to
    an AWS Direct Connect router. With this connection in place, you can
    create virtual interfaces directly to the AWS cloud (for example, to
    Amazon EC2 and Amazon S3) and to Amazon VPC, bypassing Internet
    service providers in your network path. A connection provides access
    to all AWS Regions except the China (Beijing) and (China) Ningxia
    Regions. AWS resources in the China Regions can only be accessed
    through locations associated with those Regions.
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker \email{adam.banker39@gmail.com}
}

Other contributors:
\itemize{
  \item Amazon.com, Inc. [copyright holder]
}

}
\keyword{internal}
