% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.directconnect_operations.R
\name{untag_resource}
\alias{untag_resource}
\title{Removes one or more tags from the specified AWS Direct Connect resource}
\usage{
untag_resource(resourceArn, tagKeys)
}
\arguments{
\item{resourceArn}{The Amazon Resource Name (ARN) of the resource.}

\item{tagKeys}{The tag keys of the tags to remove.}
}
\description{
Removes one or more tags from the specified AWS Direct Connect resource.
}
\examples{

}
